feat(overlay): move to helper and set default to blur
