+++ packages/react/src/helpers/index.ts
@@ -1,3 +1,4 @@
 export * from './focus-trap';
 export * from './portal';
 export * from './visually-hidden';
+export * from './overlay';
